# frontend/src/types/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RideCompass API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/axis-definitions": {
      "get": {
        "tags": [
          "axis-admin"
        ],
        "operationId": "list_axis_definitions_api_admin_axis_definitions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AxisDefinitionResponse"
                  },
                  "type": "array",
                  "title": "Response List Axis Definitions Api Admin Axis Definitions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      },
      "post": {
        "tags": [
          "axis-admin"
        ],
        "operationId": "create_axis_definition_api_admin_axis_definitions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AxisDefinitionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AxisDefinitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/admin/axis-definitions/{axis_id}": {
      "get": {
        "tags": [
          "axis-admin"
        ],
        "operationId": "get_axis_definition_api_admin_axis_definitions__axis_id__get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "axis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Axis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AxisDefinitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "axis-admin"
        ],
        "operationId": "update_axis_definition_api_admin_axis_definitions__axis_id__put",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "axis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Axis Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AxisDefinitionPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AxisDefinitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "axis-admin"
        ],
        "operationId": "delete_axis_definition_api_admin_axis_definitions__axis_id__delete",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "axis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Axis Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/axis-definitions/{axis_id}/unpublish": {
      "post": {
        "tags": [
          "axis-admin"
        ],
        "operationId": "unpublish_axis_definition_api_admin_axis_definitions__axis_id__unpublish_post",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "axis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Axis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AxisDefinitionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/axis-definitions/preview-distribution": {
      "post": {
        "tags": [
          "axis-admin"
        ],
        "operationId": "preview_axis_distribution_api_admin_axis_definitions_preview_distribution_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AxisPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValueDistribution"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/admin/axis-definitions/preview-scores": {
      "post": {
        "tags": [
          "axis-admin"
        ],
        "operationId": "preview_scores_api_admin_axis_definitions_preview_scores_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoresPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoresPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
    "/api/admin/axis-definitions/preview-display-thresholds": {
      "post": {
        "tags": [
          "axis-admin"
        ],
        "operationId": "preview_display_thresholds_api_admin_axis_definitions_preview_display_thresholds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisplayThresholdsPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayThresholdsPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/api/admin/material-catalog/coverage": {
      "get": {
        "operationId": "get_material_coverage_api_admin_material_catalog_coverage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialCoverageReport"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
    "/api/admin/derived-data/freshness": {
      "get": {
        "operationId": "get_derived_data_freshness_api_admin_derived_data_freshness_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DerivedDataFreshnessReport"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AxisDefinitionPayload": {
        "properties": {
          "axis_id": {
            "type": "string",
            "minLength": 1,
            "title": "Axis Id"
          },
          "shape": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/BreakpointLinearShape-Input"
              },
              {
                "$ref": "#/components/schemas/CategoricalShape-Input"
              }
            ],
            "title": "Shape"
          },
          "default_weight": {
            "type": "number",
            "minimum": 0.0,
            "title": "Default Weight"
          },
          "label": {
            "type": "string",
            "minLength": 1,
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "category": {
            "type": "string",
            "enum": [
              "観測",
              "推定",
              "動的"
            ],
            "title": "Category",
            "default": "推定"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published",
            "default": false
          },
          "priority_overrides": {
            "items": {
              "$ref": "#/components/schemas/PriorityCondition"
            },
            "type": "array",
            "title": "Priority Overrides"
          },
          "icon_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Id"
          },
          "time_scope": {
            "type": "string",
            "enum": [
              "always",
              "night_only"
            ],
            "title": "Time Scope",
            "default": "always"
          },
          "display_thresholds_override": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Thresholds Override"
          },
          "display_band_labels_override": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Band Labels Override"
          },
          "dedicated_way_value_layer": {
            "type": "boolean",
            "title": "Dedicated Way Value Layer",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "axis_id",
          "shape",
          "default_weight",
          "label"
        ],
        "title": "AxisDefinitionPayload"
      },
      "AxisDefinitionResponse": {
        "properties": {
          "axis_id": {
            "type": "string",
            "minLength": 1,
            "title": "Axis Id"
          },
          "shape": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/BreakpointLinearShape-Output"
              },
              {
                "$ref": "#/components/schemas/CategoricalShape-Output"
              }
            ],
            "title": "Shape"
          },
          "default_weight": {
            "type": "number",
            "minimum": 0.0,
            "title": "Default Weight"
          },
          "label": {
            "type": "string",
            "minLength": 1,
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "category": {
            "type": "string",
            "enum": [
              "観測",
              "推定",
              "動的"
            ],
            "title": "Category",
            "default": "推定"
          },
          "is_published": {
            "type": "boolean",
            "title": "Is Published",
            "default": false
          },
          "priority_overrides": {
            "items": {
              "$ref": "#/components/schemas/PriorityCondition"
            },
            "type": "array",
            "title": "Priority Overrides"
          },
          "icon_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Id"
          },
          "time_scope": {
            "type": "string",
            "enum": [
              "always",
              "night_only"
            ],
            "title": "Time Scope",
            "default": "always"
          },
          "display_thresholds_override": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Thresholds Override"
          },
          "display_band_labels_override": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Band Labels Override"
          },
          "dedicated_way_value_layer": {
            "type": "boolean",
            "title": "Dedicated Way Value Layer",
            "default": false
          },
          "display": {
            "$ref": "#/components/schemas/AxisDisplaySpec"
          },
          "weight_share_when_published": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Share When Published"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "axis_id",
          "shape",
          "default_weight",
          "label",
          "description",
          "category",
          "is_published",
          "priority_overrides",
          "icon_id",
          "time_scope",
          "display_thresholds_override",
          "display_band_labels_override",
          "dedicated_way_value_layer",
          "display",
          "weight_share_when_published"
        ],
        "title": "AxisDefinitionResponse"
      },
...
      "AxisPreviewRequest": {
        "properties": {
          "shape": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/BreakpointLinearShape-Input"
              },
              {
                "$ref": "#/components/schemas/CategoricalShape-Input"
              }
            ],
            "title": "Shape"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "shape"
        ],
        "title": "AxisPreviewRequest"
      },
...
      "DisplayThresholdsPreviewRequest": {
        "properties": {
          "axis_id": {
            "type": "string",
            "minLength": 1,
            "title": "Axis Id"
          },
          "shape": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/BreakpointLinearShape-Input"
              },
              {
                "$ref": "#/components/schemas/CategoricalShape-Input"
              }
            ],
            "title": "Shape"
          },
          "priority_overrides": {
            "items": {
              "$ref": "#/components/schemas/PriorityCondition"
            },
            "type": "array",
            "title": "Priority Overrides"
          },
          "thresholds": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "minItems": 1,
            "title": "Thresholds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "axis_id",
          "shape",
          "thresholds"
        ],
        "title": "DisplayThresholdsPreviewRequest"
      },
      "DisplayThresholdsPreviewResponse": {
        "properties": {
          "dropped_on_map": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Dropped On Map"
          },
          "bands_on_map": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Bands On Map"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dropped_on_map",
          "bands_on_map"
        ],
        "title": "DisplayThresholdsPreviewResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PriorityCondition": {
        "properties": {
          "material": {
            "type": "string",
            "minLength": 1,
            "title": "Material"
          },
          "equals": {
            "type": "string",
            "title": "Equals"
          },
          "value": {
            "type": "number",
            "title": "Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "material",
          "equals",
          "value"
        ],
        "title": "PriorityCondition"
      },
...
      "ScoresPreviewRequest": {
        "properties": {
          "shape": {
            "$ref": "#/components/schemas/BreakpointLinearShape-Input"
          },
          "xs": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Xs"
          },
          "material_values": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Material Values"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "shape"
        ],
        "title": "ScoresPreviewRequest"
      },
      "ScoresPreviewResponse": {
        "properties": {
          "scores": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Scores"
          },
          "material_points": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ScorePoint"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "title": "Material Points"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scores",
          "material_points"
        ],
        "title": "ScoresPreviewResponse"
      },
...
      "ValueDistribution": {
        "properties": {
          "sample_ways": {
            "type": "integer",
            "title": "Sample Ways"
          },
          "total_km": {
            "type": "number",
            "title": "Total Km"
          },
          "quantiles": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Quantiles"
          },
          "bins": {
            "items": {
              "prefixItems": [
                {
                  "type": "number"
                },
                {
                  "type": "number"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 3,
              "minItems": 3
            },
            "type": "array",
            "title": "Bins"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sample_ways",
          "total_km",
          "quantiles",
          "bins"
        ],
        "title": "ValueDistribution"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}